pipeline/requirements.txt
numpy
pandas
statsmodels
pmdarima
psycopg[binary]
python-dotenv
requests
xgboost
scikit-learn